Poet engine modularization
