+++ src/core/src/check_consensus.rs
@@ -1,0 +1,277 @@
+/*
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
+extern crate sawtooth_sdk;
+extern crate log;
+extern crate log4rs;
+ 
+use sawtooth_sdk::consensus::{engine::*, service::Service};
+use service::Poet2Service;
+use std::sync::mpsc::{Receiver, RecvTimeoutError};
+use std::time;
+use std::str::FromStr;
+use std::cmp;
+use serde_json;
+use std::time::Duration;
+use std::time::Instant;
+use std::collections::HashMap;
+use enclave_sgx::*;
+use consensus_state::*;
+use consensus_state_store::ConsensusStateStore;
+use poet2_util;
+use database::config;
+use database::lmdb;
+use database::{DatabaseError, CliError};
+use settings_view::Poet2SettingsView;
+use fork_resolver;
+
+const DEFAULT_BLOCK_CLAIM_LIMIT:i32 = 250;
+
+/*
+* Consensus related sanity checks to be done here
+* If all checks pass but WC < CC, forced sleep is
+* induced to sync up the clocks. Sleep duration
+* in that case would be atleast CC - WC.
+*
+*/
+
+pub fn check_consensus(
+    block: Block,
+    service: &mut Poet2Service, validator_id: &Vec<u8>, 
+    poet_pub_key: &String)
+    -> bool {
+    // 1. Validator registry check
+    // 4. Match Local Mean against the locally computed
+    // 5. Verfidy BlockDigest is a valid ECDSA of
+    //    SHA256 hash of block using OPK
+
+    //\\ 2. Signature validation using sender's PPK
+
+    if !verify_wait_certificate(block.clone(), service, &poet_pub_key){
+        return false;
+    }
+
+    // 3. k-test
+    /*if validtor_has_claimed_block_limit( service ) {
+        return false;
+    }*/
+
+    // 6. z-test
+    /*if validator_is_claiming_too_frequently {
+        return false;
+    }*/
+
+    // 7. c-test
+    let block_signer = poet2_util::to_hex_string(Vec::from(block.signer_id.clone()));
+    let validator = poet2_util::to_hex_string(validator_id.to_vec());
+    
+    if validator == block_signer && validator_is_claiming_too_early( block, service){
+        return false;
+    }
+
+    //\\ 8. Compare CC & WC
+    let chain_clock = service.get_chain_clock();
+    let wall_clock = service.get_wall_clock();
+    let wait_time:u64 = 0;//get_wait_cert_json(String::from_utf8(block.payload).unwrap()).wait_time;
+    if chain_clock + wait_time > wall_clock {
+        return false;
+    }
+    true
+}
+
+fn verify_wait_certificate(
+    _block: Block,
+    service: &mut Poet2Service,
+    poet_pub_key: &String)
+    -> bool {
+    let prev_block = service.get_block(_block.previous_id.clone()).unwrap();
+    let verify_status = service.verify_wait_certificate(&_block, &prev_block, &poet_pub_key);
+    verify_status
+}
+
+
+//k-test
+fn validtor_has_claimed_block_limit( service: &mut Poet2Service ) -> bool {
+
+    let mut block_claim_limit = DEFAULT_BLOCK_CLAIM_LIMIT;
+    let mut key_block_claim_count=9;
+    let mut    poet_public_key="abcd";
+    let mut    validator_info_signup_info_poet_public_key="abcd";
+    //  let mut key_block_claim_limit = poet_settings_view.key_block_claim_limit ;     //key
+    // need to use get_settings from service
+    let key_block_claim_limit = service.get_setting_from_head(
+        String::from("sawtooth.poet.key_block_claim_limit"));
+
+    if key_block_claim_limit != "" {
+        block_claim_limit = key_block_claim_limit.parse::<i32>().unwrap();
+    }
+
+    // let mut validator_state = self.get_validator_state();//                          //stubbed
+    // if validator_state.poet_public_key == validator_info.signup_info.poet_public_key //stubbed
+
+    if poet_public_key == validator_info_signup_info_poet_public_key     //stubbed function replaced with dummy function
+    {
+        //if validator_state.key_block_claim_count >= block_claim_limit
+        if key_block_claim_count >= block_claim_limit{
+            true }
+        else { false }
+    }
+    else{ false }
+}
+
+
+//c-test
+fn validator_is_claiming_too_early( block: Block, service: &mut Poet2Service )->bool
+{
+
+    let number_of_validators = 3_u64;
+    //    number_of_validators = (validator_registry_view.get_validators()).len();  //stubbed function
+    let total_block_claim_count = block.block_num - 1;
+    let commit_block_block_num = 0_u64;
+    //    let commit_block = block_store.get_block_by_transaction_id(validator_info.transaction_id)
+    let block_number = block.block_num;
+
+    let block_claim_delay_from_settings = service.get_setting_from_head(
+        String::from("sawtooth.poet.block_claim_delay"));
+
+    let key_block_claim_delay = block_claim_delay_from_settings.parse::<u64>().unwrap();
+    let block_claim_delay = cmp::min(key_block_claim_delay, number_of_validators - 1);
+
+    if total_block_claim_count <= block_claim_delay
+    {
+        return false;
+    }
+    // need to use get_block from service expecting block_id to have been stored
+    // along with validator info in the Poet 2 module
+	
+    let blocks_claimed_since_registration  = block_number - commit_block_block_num - 1 ;
+
+    if block_claim_delay > blocks_claimed_since_registration 
+    {
+        debug!("Failed c-test");
+        return true;
+    }
+    debug!("Passed c-test");
+    return false;
+
+}
+
+fn create_context() -> Result<lmdb::LmdbContext, CliError> {
+    let path_config = config::get_path_config();
+    let statestore_path = &path_config.data_dir.join(config::get_filename());
+
+    lmdb::LmdbContext::new(statestore_path, 1, None)
+        .map_err(|err| CliError::EnvironmentError(format!("{}", err)))
+}
+
+fn open_statestore(ctx: &lmdb::LmdbContext) -> Result<ConsensusStateStore, CliError> {
+    let statestore_db = lmdb::LmdbDatabase::new(
+        ctx,
+        &["index_consensus_state"],
+    ).map_err(|err| CliError::EnvironmentError(format!("{}", err)))?;
+
+    Ok(ConsensusStateStore::new(statestore_db))
+}
+
+pub enum ResponseMessage {
+    Ack,
+    Published,
+    Received,
+}
+
+impl FromStr for ResponseMessage {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "ack" => Ok(ResponseMessage::Ack),
+            "published" => Ok(ResponseMessage::Published),
+            "received" => Ok(ResponseMessage::Received),
+            _ => Err("Invalid message type"),
+        }
+    }
+}
+
+//z-test
+/*
+fn validator_is_claiming_too_frequently(&mut self,
+                                        validator_info: ValidatorInfo,
+                                        previous_block_id: &str,
+                                        poet_settings_view: PoetSettingsView,
+                                        population_estimate: f64,
+                                        block_cache: BlockCache,
+                                        poet_enclave_module: module) -> bool {
+
+    if self.total_block_claim_count < poet_settings_view.population_estimate_sample_size {  //totalblock count-0  pop-est-1
+        return false;
+    }
+
+    let mut population_estimate_list = VecDeque::new();
+    population_estimate_list = self._build_population_estimate_list(previous_block_id, poet_settings_view,block_cache,poet_enclave_module);
+
+    population_estimate_list.insert(ConsensusState._EstimateInfo(population_estimate, previous_block_id, validator_info.id),0);
+    //[_EstimateInfo(population_estimate=2, previous_block_id='previous_id', validator_id='validator_001_key')]
+    let mut observed_wins =0.0;
+    let mut expected_wins =0.0;
+    let mut block_count =0;
+    let mut minimum_win_count = poet_settings_view.ztest_minimum_win_count as f64; // Expecting it to be a float type value else type casting is required-----3
+    let mut maximum_win_deviation = poet_settings_view.ztest_maximum_win_deviation as f64; // Expecting it to be a float type value else type casting is required---3.075
+
+
+    for estimate_info in population_estimate_list.iter(){
+        block_count += 1; //1
+        //Float and integer addition might cause error
+        expected_wins += 1.0/estimate_info.population_estimate; //0.5    estimate_info.population_estimate----2
+
+        if estimate_info.validator_id == validator_info.id {  //validator_001_key
+            observed_wins += 1.0; //1
+            if observed_wins > minimum_win_count && observed_wins > expected_wins{ // Might be comparing float with integer value
+                let mut probability = expected_wins/block_count as f64; //Depends on the lngth of the block_count
+                let mut standard_deviation = (block_count as f64 * probability * (1.0 - probability)).sqrt();
+                let mut z_score = (observed_wins - expected_wins) / standard_deviation;
+                let mut validator_info_id: &str = validator_info.id;
+                let mut validator_info_id_start = &validator_info_id[0..8];
+                let mut validator_info_id_end: Vec<char> = validator_info_id.chars().rev().take(8).collect();
+                if z_score  > maximum_win_deviation {
+
+                    info!("Validator {} (ID={}...{}): z-test failded at depth {}, z_score={} ,expected={} , observed={}",
+                            validator_info.name,
+                            validator_info_id_start,
+                            validator_info_id_end,
+                            block_count,
+                            z_score,
+                            expected_wins,
+                            observed_wins);
+
+                    return true;
+                }
+            }
+        }
+    }
+    let validator_info_id = validator_info.id;
+    let validator_info_id_start = &validator_info_id[0..8];
+    let mut validator_info_id_end: Vec<char> = validator_info_id.chars().rev().take(8).collect();
+    info!("Validator {} (ID={}...{}): zTest succeeded at depth {}, expected={} , observed={}",
+                            validator_info.name,
+                            validator_info_id_start,
+                            validator_info_id_end,
+                            block_count,
+                            expected_wins,
+                            observed_wins);
+
+    return false;
+}*/
